Description
task for this assignment is to use C++ language to implement a net pay function. 
 
1.	Implement a net pay function using the C++ programming language.  The program should begin by obtaining three input values for hours worked, rate of pay and deductions.  Net pay is determined by computing gross pay and then subtracting deductions.  Gross pay is the sum of hours worked up to 40 multiplied by rate of pay and hours worked over 40 multiplied by 1.5 times rate of pay, i.e. overtime. 

A function called “netpay” should receive as arguments the three input values and should return the computed net pay as a result.  The program should be run two times, once for a case of overtime and once for a case of no overtime.

2.	The program should output the input values and the computed net pay, all clearly labeled for the reader. The program is to be run 2 separate times. Each program should should produce an output text file named csc231_yourname_program2_overtime.txt or csc231_yourname_program2_noovertime.txt depending on the calculation.  The following information on separate lines should be at the top of the report files:

a.	The course name, ID and section
b.	your name
c.	the file name
d.	the program assignment number and due date
e.	the program purpose

3.	The program file should be named csc231_lastname_prog2.cpp. Your C++ program should contain comments starting on line 1 containing information similar to 2a through 2e above.  You are encouraged to add additional comments throughout the program that your feel might be helpful to the reader of your source code.